requirements.txt
BeautifulSoup==3.2.1
Django==1.7b2
Pillow==2.4.0
SPARQLWrapper==1.5.2
South==0.8.2
boto==2.8.0
bottlenose==0.4.2
django-appconf==0.6
django-blogango==0.6.2
django-classy-tags==0.4
django-debug-toolbar==1.2
django-haystack==2.0.0
django-markupfield==1.0.2
django-pingback==0.3
-e git://github.com/partengine/django-pure-pagination@9265315c7c27bc4d870547a419d99ef7f009ccff#egg=django_pure_pagination-master
django-registration==1.0
django-storages==1.1.8
django-taggit==0.10
django-tastypie==0.11.0
django-xmlrpc==0.1.4
feedparser==5.1.3
html5lib==1.0b3
isodate==0.4.9
lxml==3.2.1
markdown2==2.1.0
microdata==0.4.0
mimeparse==0.1.3
psycopg2==2.4.6
pyelasticsearch==0.6
pyparsing==1.5.7
python-amazon-simple-product-api==1.4.0
python-dateutil==2.2
python-memcached==1.48
python-mimeparse==0.1.4
rdflib==4.0.1
requests==1.2.3
simplejson==3.0.8
six==1.6.1
sorl-thumbnail==11.12.1b
sqlparse==0.1.11
wsgiref==0.1.2
